requirements.txt
asgiref==3.6.0
cfgv==3.3.1
distlib==0.3.6
Django==4.1.7
django-extensions==3.2.1
djangorestframework==3.14.0
filelock==3.9.0
identify==2.5.18
nodeenv==1.7.0
platformdirs==3.0.0
pre-commit==3.1.0
pydotplus==2.0.2
pyparsing==3.0.9
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
virtualenv==20.19.0
